New configuration functions added.
There's a new function to parse pacman'sm configuration file (this function does not initialize the library, its job is to fill a structure for later use).
One function to initilize the alpm library (this function uses the structure to fill some mandatory options and then it initilize the library).
Another function to release all the resources used by the library.