Some bug were fixed and implementation of the UI.
Some bugs were fixed in the src/Makefile.am (including the addition of the new files) and in the src/conf.c, new user interface was added.
the main.c file has been renamed to aum.c, and a new header file aum.h. This files it's going to be use for global definitions.