Minor mofications to conf.c
There're some directives in pacman's configuration that need to be store on a list, this directives are: IgnorePkg, IgnoreGrps, NoUpgrade, NoExtract.

Since we're using the same data structure that alpm is using (alpm_list_t *), there's no need to call alpm_option_add_xxxx (where xxxx is one of the directives describe above) to add these options to the library's structure, instead we'll use alpm_option_set_xxxx to use the same list.
This is a better design because we save up some memory.